Fix circular dependency by moving out type
