+++ src/components/file-system-explorer/types.ts
@@ -1,0 +1,19 @@
+export interface ActiveFile {
+  type: string;
+  href: string;
+  zipState: string;
+  attributes: {
+    accessTime: string;
+    dir: boolean;
+    group: string;
+    id: number;
+    modificationTime: string;
+    name: string;
+    owner: string;
+    parentId: number;
+    path: string;
+    permission: string;
+    size: number;
+    underConstruction: boolean;
+  };
+}
